Update Demo app
