+++ demo-app/src/main/java/com/gravatar/demoapp/ui/DemoProfileSummaryCard.kt
@@ -7,19 +7,11 @@ import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.neverEqualPolicy
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.gravatar.restapi.models.Profile
-import com.gravatar.services.GravatarResult
-import com.gravatar.services.ProfileService
 import com.gravatar.types.Email
 import com.gravatar.ui.GravatarTheme
 import com.gravatar.ui.components.ComponentState
@@ -27,26 +19,12 @@ import com.gravatar.ui.components.ProfileSummary
 import com.gravatar.ui.components.atomic.Avatar
 
 @Composable
-fun DemoProfileSummaryCard(email: String, avatarCache: String? = null, modifier: Modifier = Modifier) {
-    val profileService = ProfileService()
-
-    var profileState: ComponentState<Profile> by remember { mutableStateOf(ComponentState.Loading, neverEqualPolicy()) }
-
-    LaunchedEffect(email) {
-        profileState = ComponentState.Loading
-        when (val result = profileService.retrieveCatching(Email(email))) {
-            is GravatarResult.Success -> {
-                result.value.let {
-                    profileState = ComponentState.Loaded(it)
-                }
-            }
-
-            is GravatarResult.Failure -> {
-                profileState = ComponentState.Empty
-            }
-        }
-    }
-
+fun DemoProfileSummaryCard(
+    email: String,
+    profileState: ComponentState<Profile> = ComponentState.Loading,
+    avatarCache: String? = null,
+    modifier: Modifier = Modifier,
+) {
     Card(
         modifier = modifier.fillMaxWidth(),
         elevation = CardDefaults.cardElevation(defaultElevation = 1.dp),
